+++ src/events/events_block_comment_change.ts
@@ -1,0 +1,67 @@
+/**
+ * @license
+ * Copyright 2024 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import * as Blockly from "blockly/core";
+import {
+  BlockCommentBase,
+  BlockCommentBaseJson,
+} from "./events_block_comment_base";
+import type { ScratchCommentBubble } from "../scratch_comment_bubble";
+
+class BlockCommentChange extends BlockCommentBase {
+  oldContents_: string;
+  newContents_: string;
+
+  constructor(
+    opt_blockComment?: ScratchCommentBubble,
+    oldContents?: string,
+    newContents?: string
+  ) {
+    super(opt_blockComment);
+    this.type = "block_comment_change";
+    this.oldContents_ = oldContents;
+    this.newContents_ = newContents;
+    // Disable undo because Blockly already tracks changes to comment text for
+    // undo purposes; this event exists solely to keep the Scratch VM apprised
+    // of the state of things.
+    this.recordUndo = false;
+  }
+
+  toJson(): BlockCommentChangeJson {
+    return {
+      ...super.toJson(),
+      newContents: this.newContents_,
+      oldContents: this.oldContents_,
+    };
+  }
+
+  static fromJson(
+    json: BlockCommentChangeJson,
+    workspace: Blockly.Workspace,
+    event?: any
+  ): BlockCommentChange {
+    const newEvent = super.fromJson(
+      json,
+      workspace,
+      event ?? new BlockCommentChange()
+    ) as BlockCommentChange;
+    newEvent.newContents_ = json["newContents"];
+    newEvent.oldContents_ = json["oldContents"];
+
+    return newEvent;
+  }
+}
+
+interface BlockCommentChangeJson extends BlockCommentBaseJson {
+  newContents: string;
+  oldContents: string;
+}
+
+Blockly.registry.register(
+  Blockly.registry.Type.EVENT,
+  "block_comment_change",
+  BlockCommentChange
+);